Separate out app and root
